pyomq: extract PeerInfo to fix blake3zmq-without-curve compilation
- Move `PeerInfo` pyclass from `auth.rs` into `peer_info.rs`, gated on
  `any(curve, blake3zmq)`. Fixes compilation with `--features blake3zmq`
  alone, where `crate::auth` was absent (`mod auth` requires `curve`).
- Fix missing `park_end()` in `send_message` slow path when socket is
  closed mid-send. Matches the recv path's discipline.
- Fix stale doc comment referencing "compio thread" after the tokio port.
