+++ bindings/pyomq/src/auth.rs
@@ -1,43 +1,13 @@
-//! CURVE client authentication: `PeerInfo` pyclass and authenticator
-//! bridge from Python callables / key lists to `omq_proto::Authenticator`.
+//! CURVE client authentication: authenticator bridge from Python
+//! callables / key lists to `omq_proto::Authenticator`.
 
 use std::collections::HashSet;
 use std::fmt;
 
 use pyo3::prelude::*;
-use pyo3::types::PyBytes;
 
+use crate::peer_info::PeerInfo;
 use crate::socket::SocketInner;
-
-/// Peer information passed to a Python authenticator callback.
-#[pyclass(frozen, module = "pyomq._native")]
-pub struct PeerInfo {
-    public_key: Py<PyBytes>,
-}
-
-#[pymethods]
-impl PeerInfo {
-    #[getter]
-    fn public_key(&self, py: Python<'_>) -> Py<PyBytes> {
-        self.public_key.clone_ref(py)
-    }
-}
-
-impl PeerInfo {
-    fn from_raw(py: Python<'_>, raw: &[u8; 32]) -> Self {
-        let pk = omq_proto::CurvePublicKey::from_bytes(*raw);
-        let z85 = pk.to_z85();
-        Self {
-            public_key: PyBytes::new_bound(py, z85.as_bytes()).unbind(),
-        }
-    }
-
-    pub(crate) fn from_raw_bytes(py: Python<'_>, raw: &[u8; 32]) -> Self {
-        Self {
-            public_key: PyBytes::new_bound(py, raw).unbind(),
-        }
-    }
-}
 
 /// Two modes of CURVE client authentication stored in the Overlay
 /// before socket materialization.
